backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.3.56
numpy==1.21.2
SpeechRecognition==3.10.1
Werkzeug==3.0.1
transformers==4.11.3
torch==1.9.0
Pillow==8.3.2 